2.2.0 Release 1 - Early Build
An early build for 2.2.0.
Release: 1
Early Build

+++ index.js
@@ -1,17 +1,23 @@
 const { CommandoClient } = require("discord.js-commando");
-const BotConfiguration = require("./BotConfiguration.js");
-const BadWords = require("./BadWords.js");
-const BotData = require("./BotData.js");
+const BotConfiguration = require("./Configuration.js");
+const System = require("./System.js");
 const discord = require("discord.js");
 const token = require("./Token.js");
 const chalk = require("chalk");
 const db = require("quick.db");
 const path = require("path");
+const fs = require("fs");
 const ms = require("ms");
 
 const bot = new CommandoClient({
 	commandPrefix: BotPrefix,
+    owner: "453719594708762636", //XXX Temp?
 });
+
+//TODO Check each file for correct class
+//TODO Have it say Currently Muted in -stats command
+//TODO Add edited messages auto moderation
+//TODO Auto Moderation | Edited Messages
 
 bot.registry
 	.registerDefaultTypes()
@@ -36,15 +42,8 @@ bot.on('ready', function(){
     console.log(`Running Version: ${Version}`);
 });
 
-if (db.get("settings.StaffApplications")== null)db.add("settings.StaffApplications", StaffApplicationsSetting);
-if (db.get("settings.AutoModeration")== null)db.add("settings.AutoModeration", AutoModerationSetting);
-if (db.get("settings.DeadChatPings")== null)db.add("settings.DeadChatPings", DeadChatPingSetting);
-if (db.get("settings.LevelUpSystem")== null)db.add("settings.LevelUpSystem", LevelUpsSetting);
-
-//----------------------------
-
-//New Members
-bot.on("guildMemberAdd", member => {
+//New Member Message
+bot.on('guildMemberAdd', member => {
     //Welcome Message
     const NewMemberMessage = new discord.MessageEmbed()
         .setTimestamp()
@@ -78,10 +77,6 @@ bot.on('guildMemberRemove', member => {
         `)
     let MemberLeaveChannel = member.guild.channels.cache.get(MemberLeaveChannelID);
     MemberLeaveChannel.send(MemberLeaveMessage);
-
-    member.send("We're sorry to see you go. If you don't mind, can you tell us why you left? https://forms.gle/UbyAV2Nze9ni24mx5").catch(err => {
-        console.log(`Could not message the person who left!`);
-    });
 });
 
 bot.on('message', function(message){
@@ -98,17 +93,20 @@ bot.on('message', function(message){
     if (message.content == "Bob Bingi"){
         message.channel.send("It's the one and only Bob Bingi! Introducing Markiplier and B.B! https://www.youtube.com/watch?v=0Pocn8aSWS4 make sure to watch!");
     }
-    var MCIPListener = [" ip ", " adress "];
+    if (message.content == "dapopo"){
+        message.reply("You knew him?");
+    }
+    //Listener for Minecraft IP
+    var MCIPListener = [" ip ", " address "];
     let msg = message.content.toLowerCase();
     for (x = 0; x < MCIPListener.length; x++){
         if (msg.includes(MCIPListener[x])){
             if (message.author.bot)return;
-            message.reply("If you need the ip for the Minecraft server, please visit <#704460219521957907>!");
+            message.reply("If you need the IP for the Minecraft server, please visit <#704460219521957907>!");
         }
     }
     //Level Up System
     if (db.get("settings.LevelUpSystem")== 1){
-        //XP Generator
         if (message.author.bot)return;
         if (message.guild === null)return;
         if (message.channel.id === `${ChannelExcludeID}`)return;
@@ -132,7 +130,7 @@ bot.on('message', function(message){
                     **User:** ${message.author}
                     **Level:** ${db.get(`${message.author.id}.basic.level`)}
                 `)
-                .setFooter(`You have recieved $${LevelUpMoney}! Nice job!`)
+                .setFooter(`You have received $${LevelUpMoney}! Nice job!`)
             let LevelUpChannel = message.guild.channels.cache.get(LevelUpChannelID);
             LevelUpChannel.send(LevelUpMessage);
         }
@@ -146,38 +144,36 @@ bot.on('message', function(message){
         if (message.author.bot)return;
         let ModLogChannel = message.guild.channels.cache.get(ModLogID);
         //Mute Bypass Protection
-        if (MuteBypassProtectionSetting == "1"){
-            if (db.get(`${message.author.id}.admin.CurrentlyMuted`)== 1){
-                message.delete();
-                db.add(`${message.author.id}.admin.TimesBypassedMute`, 1);
-                let MuteRole = message.guild.roles.cache.get(MuteRoleID);
-                message.member.roles.add(MuteRole);
+        if (MuteBypassProtectionSetting == "1" && db.get(`${message.author.id}.admin.CurrentlyMuted`)== 1){
+            message.delete();
+            db.add(`${message.author.id}.admin.TimesBypassedMute`, 1);
+            let MuteRole = message.guild.roles.cache.get(MuteRoleID);
+            message.member.roles.add(MuteRole);
 
-                //Chat Response
-                const MuteBypassMessage = new discord.MessageEmbed()
-                    .setTimestamp()
-                    .setColor("")//Keep Empty
-                    .setAuthor(message.author.tag, message.author.displayAvatarURL())
-                    .setDescription(`
-                        **Mute Bypasses:** ${db.get(`${message.author.id}.admin.TimesBypassedMute`)}
-                    `)
-                    .setFooter("Auto Moderation: Mute Bypass")
-                message.channel.send(MuteBypassMessage).then(message => {
-                    message.delete({timeout: 15000});
-                });
+            //Chat Response
+            const MuteBypassMessage = new discord.MessageEmbed()
+                .setTimestamp()
+                .setColor("")//Keep Empty
+                .setAuthor(message.author.tag, message.author.displayAvatarURL())
+                .setDescription(`
+                    **Mute Bypasses:** ${db.get(`${message.author.id}.admin.TimesBypassedMute`)}
+                `)
+                .setFooter("Auto Moderation: Mute Bypass")
+            message.channel.send(MuteBypassMessage).then(message => {
+                message.delete({timeout: 15000});
+            });
 
-                //Logged Resonse
-                const MuteBypassMessageLog = new discord.MessageEmbed()
-                    .setTimestamp()
-                    .setColor("")//Keep Empty
-                    .setAuthor("Mute Bypass | "+message.author.tag, message.author.displayAvatarURL())
-                    .setDescription(`
-                        **User:** ${message.author}
-                        **Bypasses:** ${db.get(`${message.author.id}.admin.TimesBypassedMute`)}
-                    `)
-                    .setFooter("Auto Moderation: Mute Bypass")
-                ModLogChannel.send(MuteBypassMessageLog);
-            }
+            //Logged Response
+            const MuteBypassMessageLog = new discord.MessageEmbed()
+                .setTimestamp()
+                .setColor("")//Keep Empty
+                .setAuthor("Mute Bypass | "+message.author.tag, message.author.displayAvatarURL())
+                .setDescription(`
+                    **User:** ${message.author}
+                    **Bypasses:** ${db.get(`${message.author.id}.admin.TimesBypassedMute`)}
+                `)
+                .setFooter("Auto Moderation: Mute Bypass")
+            ModLogChannel.send(MuteBypassMessageLog);
         }else{
             return;
         }
@@ -298,46 +294,14 @@ bot.on('messageDelete', async (message) => {
 bot.on('ready', () => {
     setInterval(() => {
         if (db.get("settings.DeadChatPings") == 1){
-            var DeadChatQuestion = Math.round(Math.random() * 31);
-            if (DeadChatQuestion == 7){DCPQuestion = "What would you do if you where offered the chance to go to Mars?"};
-            if (DeadChatQuestion == 0){DCPQuestion = "What is the most valuable thing you currently have ingame?"};
-            if (DeadChatQuestion == 13){DCPQuestion = "What is one of the funniest (but clean) jokes you know?"};
-            if (DeadChatQuestion == 1){DCPQuestion = "What movie or book character do you most identify with?"};
-            if (DeadChatQuestion == 2){DCPQuestion = "As a child, what did you wish to be when you grew up?"};
-            if (DeadChatQuestion == 3){DCPQuestion = "Are we seeing signs of evolution in our species?"};
-            if (DeadChatQuestion == 4){DCPQuestion = "What's a trait do you like most about yourself?"};
-            if (DeadChatQuestion == 6){DCPQuestion = "Why is science so important to modern society?"};
-            if (DeadChatQuestion == 8){DCPQuestion = "What is your favorite form of transportation?"};
-            if (DeadChatQuestion == 10){DCPQuestion = "What is your favorite version of Minecraft?"};
-            if (DeadChatQuestion == 11){DCPQuestion = "Is time relative to a person or universal?"};
-            if (DeadChatQuestion == 5){DCPQuestion = "What subject are you the best at in school?"};
-            if (DeadChatQuestion == 12){DCPQuestion = "What song always puts you in a good mood?"};
-            if (DeadChatQuestion == 14){DCPQuestion = "Survival, Creative or Hardcore Minecraft?"};
-            if (DeadChatQuestion == 15){DCPQuestion = "What do you like to do on the weekends?"};
-            if (DeadChatQuestion == 16){DCPQuestion = "Would you say you make friends easily?"};
-            if (DeadChatQuestion == 17){DCPQuestion = "What do you like to do on a rainy day?"};
-            if (DeadChatQuestion == 18){DCPQuestion = "What's your favourite type of music?"};
-            if (DeadChatQuestion == 19){DCPQuestion = "What is your favourite Disney movie?"};
-            if (DeadChatQuestion == 9){DCPQuestion = "Have you read the #qod of the day?"};
-            if (DeadChatQuestion == 20){DCPQuestion = "What's your favorite activity?"};
-            if (DeadChatQuestion == 21){DCPQuestion = "Laptop, Desktop or Handheld?"};
-            if (DeadChatQuestion == 22){DCPQuestion = "What's your favorite food?"};
-            if (DeadChatQuestion == 23){DCPQuestion = "Java or Bedrock Minecraft?"};
-            if (DeadChatQuestion == 24){DCPQuestion = "Which is the better pie? Debate in chat!"};
-            if (DeadChatQuestion == 25){DCPQuestion = "Windows, MacOS or Linux?"};
-            if (DeadChatQuestion == 26){DCPQuestion = "Playstation Or Xbox?"};
-            if (DeadChatQuestion == 27){DCPQuestion = "How have you been?"};
-            if (DeadChatQuestion == 28){DCPQuestion = "Iphone or Android?"};
-            if (DeadChatQuestion == 29){DCPQuestion = "Do you have pets?"};
-            if (DeadChatQuestion == 30){DCPQuestion = "Airplane or Car?"};
-
+            const DCPQuestion = Math.floor(Math.random() * DeadChatQuestions.length);
             const DeadChatMessage = new discord.MessageEmbed()
                 .setTimestamp()
                 .setColor("RANDOM")
                 .attachFiles('./Images/DCP_Question.png')
                 .setThumbnail('attachment://DCP_Question.png')
                 .setTitle("Dead Chat Ping!")
-                .addField(DCPQuestion, `<@&${DCPPingRoleID}>`)
+                .addField(DeadChatQuestions[DCPQuestion], `<@&${DCPRoleID}>`)
             let PingChannel = bot.channels.cache.get(DCPChannelID);
             PingChannel.send(DeadChatMessage);
 
@@ -349,5 +313,5 @@ bot.on('ready', () => {
         }else{
             return;
         }
-    }, 1000 * 60 * 60 * PingTime);
+    }, ms(`${PingTime}`));
 });
